Move colors package to data.colors
This way the java-chartjs-data module is the .data package
