+++ data/src/main/java/nl/crashdata/chartjs/data/ChartJsDataset.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import nl.crashdata.chartjs.colors.ChartJsRGBAColor;
+import nl.crashdata.chartjs.data.colors.ChartJsRGBAColor;
 
 /**
  * Represents the configuration of one dataset for a chart. In other words, this represents the data
