+++ data/src/main/java/nl/crashdata/chartjs/data/simple/builder/SimpleChartJsDatasetBuilder.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import nl.crashdata.chartjs.colors.ChartJsRGBAColor;
 import nl.crashdata.chartjs.data.ChartJsFill;
+import nl.crashdata.chartjs.data.colors.ChartJsRGBAColor;
 import nl.crashdata.chartjs.data.simple.SimpleChartJsDataset;
 
 public class SimpleChartJsDatasetBuilder<E extends Serializable>
